Fixed file not found bug

+++ generate_thought_vectors.py
@@ -23,7 +23,8 @@ def main():
 	model = skipthoughts.load_model()
 	caption_vectors = skipthoughts.encode(model, captions)
 
-	os.remove(join(args.data_dir, 'sample_caption_vectors.hdf5'))
+	if os.path.isfile(join(args.data_dir, 'sample_caption_vectors.hdf5')):
+		os.remove(join(args.data_dir, 'sample_caption_vectors.hdf5'))
 	h = h5py.File(join(args.data_dir, 'sample_caption_vectors.hdf5'))
 	h.create_dataset('vectors', data=caption_vectors)		
 	h.close()
